pyproject.toml
[project]
name = "voicecraft"
version = "0.1.0"
description = "AI-powered speech synthesis tool with multi-speaker support for OpenAI and Gemini models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.106.1",
    "google-genai>=0.8.0",
    "litellm>=1.50.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
voicecraft = "voicecraft.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"